consolidate keyboard configs (DayKev suggestion)
